// src/api/openapi/auto-generated/models/LocationAnalysisData.ts
export interface LocationAnalysisData {
    /**
     * Internal system location ID.
     * @type {string}
     * @memberof LocationAnalysisData
     */
    locationId: string;
    /**
     * Customer-provided location ID.
     * @type {number}
     * @memberof LocationAnalysisData
     */
    customerLocationId: number;
    /**
     * @type {number}
     * @memberof LocationAnalysisData
     */
    adaptedAal2025: number;
    /**
     * @type {number}
     * @memberof LocationAnalysisData
     */
    unadaptedAal2025: number;
    /**
     * Customer-provided location name.
     * @type {string}
     * @memberof LocationAnalysisData
     */
    name?: string;
    /**
     * 
     * @type {number}
     * @memberof LocationAnalysisData
     */
    capex?: number;
    /**
     * 
     * @type {number}
     * @memberof LocationAnalysisData
     */
    growthCapex?: number;
    /**
     * 
     * @type {number}
     * @memberof LocationAnalysisData
     */
    implementationYear?: number;
    /**
     * 
     * @type {number}
     * @memberof LocationAnalysisData
     */
    netPresentValue?: number;
    /**
     * 
     * @type {number}
     * @memberof LocationAnalysisData
     */
    returnOnInvestment?: number;
    /**
     * 
     * @type {number}
     * @memberof LocationAnalysisData
     */
    breakevenYear?: number;
    /**
     *
     * @type {string}
     * @memberof LocationAnalysisData
     */
    occupancyName?: string;
    /**
     * If there was an error during financial computation for the location, this will be populated. 
     * @type {string}
     * @memberof LocationAnalysisData
     */
    financialError?: string;
    /**
     * A map of a group to its strategies and their values. This structure is TBD, and may change to accommodate e.g. applicability.
     * @type {{ [key: string]: GroupStrategySettings; }}
     * @memberof LocationAnalysisData
     */
    strategies: { [key: string]: GroupStrategySettings; };
}